fix(keystore): component restructing
